# Optimize writing RLE runs in parquet column descriptors