--- a/lib/trino-parquet/src/test/java/io/trino/parquet/BenchmarkParquetFormat.java
+++ b/lib/trino-parquet/src/test/java/io/trino/parquet/BenchmarkParquetFormat.java
@@ -100,8 +100,6 @@ public class BenchmarkParquetFormat
             throws IOException
     {
         data = dataSet.createTestData();
-        targetDir.mkdirs();
-        writeData(new File(targetDir, UUID.randomUUID().toString()));
     }
 
     @TearDown